Add itemId and refactor
